<feature> add podcast episode count to site nav and learn the difference between data and collections
The feature was actually easy, but I broke the podcasts for a while while
I learned how things worked.

I need to remember that COLLECTIONS data is only what is in the front matter
of the pagination page that creates the template.  All the podcast feed data
does NOT automatically move over.

I have to include what I need in the front matter (Eleventy Computed) or reference it as
global data.  I chose to reference it as global data.