<feature> add tag counts and sort tag page by post count
Created a new collection to count all occurences of a tag and sort the list in descending order.

The tag page now uses this new collection to display the tags that can be browsed through.

NOTE:  It still includes the 'Posts' tag, which is a catch all.  Either I need to filter it out or get rid of the posts category all together.